Forse qualche dio non ha finito con noi

+++ azzeroCO2-war/src/main/java/it/agilis/mens/azzeroCO2/shared/model/conoscoCO2/ConoscoCO2Model.java
@@ -11,12 +11,11 @@ import com.extjs.gxt.ui.client.data.BaseModel;
  */
 public class ConoscoCO2Model extends BaseModel {
 
-    public void setConoscoCO2(Integer conoscoCO2) {
+    public void setConoscoCO2(Double conoscoCO2) {
         set("conoscoCO2", conoscoCO2);
     }
 
     public Double getConoscoCO2() {
         return get("conoscoCO2");
-
     }
 }
